refactor(website): move mdx component config to separate file
